backend/requirements.txt
# EMO Backend - Requirements
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0


# OpenAI-compatible client (for LitGPT / Local LLMs)
openai>=1.0.0
langchain-openai>=0.1.0

# Google APIs (Gmail)
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.100.0

# Document parsing
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.0.0
pypdf>=4.0.0

# Vector database
chromadb>=0.5.0

# LangGraph / LangChain (for Emo2)
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0

# Web & YouTube tools
requests>=2.31.0
youtube_transcript_api>=0.6.0
beautifulsoup4>=4.12.0
